fix after rewiev

+++ controllers/movie.js
@@ -5,7 +5,8 @@ const ForbiddenError = require('../errors/forbidden-error');
 const ValidationError = require('../errors/validation-err');
 
 const getMovies = (req, res, next) => {
-  Movie.find({}, { __v: 0, _id: 0 })
+  const { _id } = req.user;
+  Movie.find({ owner: _id }, { __v: 0, _id: 0, owner: 0 })
     .then((movies) => {
       if (!movies) throw new ServerError();
       res.status(200).send({ data: movies });
@@ -14,8 +15,34 @@ const getMovies = (req, res, next) => {
 };
 
 const createMovie = (req, res, next) => {
-  const { name, link } = req.body;
-  Movie.create({ name, link, owner: req.user._id }, { __v: 0, _id: 0 })
+  const { _id } = req.user;
+  const {
+    nameRU,
+    nameEN,
+    thumbnail,
+    trailer,
+    image,
+    description,
+    year,
+    duration,
+    director,
+    country,
+    movieId,
+  } = req.body;
+  Movie.create({
+    nameRU,
+    nameEN,
+    thumbnail,
+    trailer,
+    image,
+    description,
+    year,
+    duration,
+    director,
+    country,
+    movieId,
+    owner: _id,
+  }, { __v: 0, _id: 0, owner: 0 })
     .then((movie) => res.status(200).send({ data: movie }))
     .catch((err) => {
       if (err.name === 'ValidationError') throw new ValidationError({ message: 'Неправильно введены данные' });
@@ -24,13 +51,15 @@ const createMovie = (req, res, next) => {
 };
 
 const deleteMovie = (req, res, next) => {
-  Movie.findById(req.params.movieId)
+  const { movieId } = req.params;
+  const { _id } = req.user;
+  Movie.findById(movieId)
     .then((movie) => {
       if (!movie) throw new NotFoundError('Фильм не найден');
-      if (String(movie.owner) !== req.user._id) throw new ForbiddenError('Вы не можете удалить этот фильм');
+      if (String(movie.owner) !== _id) throw new ForbiddenError('Вы не можете удалить этот фильм');
       return movie;
     }).then((movie) => {
-      Movie.findByIdAndDelete(movie.id, { __v: 0, _id: 0 })
+      Movie.findByIdAndDelete(movie.id, { __v: 0, _id: 0, owner: 0 })
         .then(() => res.status(200).send({ data: movie }));
     })
     .catch(next);
